Exclude libpng and lijpeg from build.
On some machines, there is a symbol conflict between Tensorflow and OpenCV, most likely due to Tensorflow wrongly exposing some parts of libpng or libjpeg. This results is cv::imread not being able to read png or jpeg files.
To avoid such a conflict, we exclude libpng and libjpeg from the build. This however prevents some ops, such as image encoding or decoding, from being used, but we do not consider these as critical when deploying Tensorflow models in real applications.